requirements.txt
flask
flask_restful
webargs
requests
python-dotenv
